backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
requests==2.32.5
python-dotenv==1.0.1
pydantic==2.11.10